Update README.md
Building world-class tech teams just got easier!

Korio is on mission to spur job creation on the African continent by connecting Africa top tech talents with global jobs. We offer IT and digital outsourcing solutions to startups and global companies. Our goal is to enhance efficiency and deliver seamless digital transformation for our clients, while connecting them with the best tech teams for mutual wins.

Evidence suggest that small investments that connect youth to jobs and increase their human capital yield greater impacts on the African continent. We believe that cultivating a skilled workforce is the formula to spur job creation in Africa, while delivering exceptional expertise globally. As such, we are nurturing the next generation of developers and digital marketers through providing talented young people on the continent with tech-based learning experiences that equip them for global labor markets.

Working with Korio becomes not only a smart business decision, but an impact investment opportunity for our clients, as we contribute to enable a sustainable society, where talented young people in Africa have access to equal employment opportunities globally.